+++ javaScript/projects/weather-app/src/helpers/formatTime.js
@@ -1,0 +1,12 @@
+export default function formatTime(time) {
+  // returns 12 hour time format
+  // ex: 2:00 PM or 10:00 AM
+  const date = new Date(`12-12-2000 ${time}`);
+  const timeProperties = {
+    hour: 'numeric',
+    minute: 'numeric',
+    hour12: true,
+  };
+
+  return date.toLocaleString('en-us', timeProperties);
+}